requirements.txt
amqp==2.6.0
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.1
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
cssselect==1.1.0
Django==3.0.8
djangorestframework==3.11.0
fake-useragent==0.1.11
future==0.18.2
idna==2.10
kombu==4.6.11
lxml==4.5.2
parse==1.15.0
pyee==7.0.2
pyppeteer==0.2.2
pyquery==1.4.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-html==0.10.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tqdm==4.48.0
urllib3==1.25.9
vine==1.3.0
w3lib==1.22.0
websockets==8.1
